Add single ErrorsModerate component
