+++ api/schema/req.body.ErrorModerate.json
@@ -1,14 +1,10 @@
 {
     "type": "object",
     "required": [
-        "job",
         "moderate"
     ],
     "additionalProperties": false,
     "properties": {
-        "job": {
-            "type": "integer"
-        },
         "moderate": {
             "type": "string",
             "enum": [
